Updated osx dylibs to use @executable_path in their library linkages
Rather than libgdiplus.dylib referencing libpng14.14.dylib in
the /Library/Frameworks/Mono.frakwork directory it should be
looking for it in the same directory.

You can use the install_name_tool to change this information.
This should be done if the files are upgraded.

To change path of a link to another library

	install_name_tool -change @executable_path/libpng14.14.dylib /Library/Frameworks/Mono.framework/Versions/4.2.1/lib/libpng14.14.dylib libfreetype.6.dylib

To change id of actual library

	install_name_tool -id "@loader_path/libfreetype.6.dylib" libfreetype.6.dylib

To list dependencies

	otool -L libfreetype.6.dylib