Fix an issue running mcs when running under Steam
Steam sets the LD_PRELOAD environment variable to
force the loading of its overlay early on in the
load process.

The problem is they attempt to load both 32 and 64
bit overlays. Probably because they don't know if
the game is 32 or 64 bit. This results in `ld`
issuing errors (which it ignores) on the stderr.

When `mcs` is called by mono to compile a temp
assembly during serialization it checks the
stderr for issues. Even though the `ld` errors
are not related to `mcs` it still fails. The
result is that the serialization system thinks
the build failed.

This fix we clear out the LD_PRELOAD once we
start the app. We know they will have been loaded
already and this prevents `mcs` trying to load them
too.